added redirect in admin pages

+++ src/app/admin/users/components/UserRow.tsx
@@ -1,31 +1,34 @@
 "use client";
 
+import React from "react";
 import Link from "next/link";
-import { UserType } from "../../../../types";
-import { deleteUser } from "../actions/deleteUser";
 import { useRouter } from "next/navigation";
+import { deleteUser } from "../actions/deleteUser";
+import { UserType } from "@/types";
 
-export default function UserRow({ user }: { user: UserType }) {
-  const router = useRouter();
+interface UserRowProps {
+  user: UserType;
+  refresh: () => void;
+}
 
+export default function UserRow({ user, refresh }: UserRowProps) {
   const handleDelete = async () => {
+    const confirmed = confirm(`Are you sure you want to delete "${user.username}"?`);
+    if (!confirmed) return;
+
     await deleteUser(user._id);
-    router.refresh();
+    refresh();
   };
 
   return (
     <tr>
-  <td className="px-6 py-4">{user.username}</td>  
-      <td>{user.email}</td>
-      <td>
-        
-        <Link
-          href={`/admin/users/edit/${user._id}`}
-          className="text-blue-500 mr-4"
-        >
-          Edit
+      <td className="border px-4 py-2">{user.username}</td>
+      <td className="border px-4 py-2">{user.email}</td>
+      <td className="border px-4 py-2 space-x-2">
+        <Link href={`/admin/users/edit/${user._id}`}>
+          <button className="text-blue-600 hover:underline">Edit</button>
         </Link>
-        <button onClick={handleDelete} className="text-red-500">
+        <button onClick={handleDelete} className="text-red-600 hover:underline">
           Delete
         </button>
       </td>
